Limit Target fields size and form a bit
